refactor: remove prepare command
